% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_AC_mn}
\alias{cciea_AC_mn}
\title{Monthly Mean Crab Landings 1990-2015}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{month}{ () []}
\item{month_number}{ () [1, 12]}
\item{landings}{Landings (monthly mean) 1990-2015 (tons) [0.1127933, 356.5503]}
\item{fish_tickets}{Fish Tickets (monthly mean0 1990-2015 (counts) [1.222222, 326.8811]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_AC_mn/index.html}
}
\usage{
cciea_AC_mn
}
\description{
Pending
}
\concept{dataset_erddap}
\keyword{datasets}
