% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_B_AS_DIET}
\alias{cciea_B_AS_DIET}
\title{Seabird diet}
\format{
A data frame with 669 rows and 3 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{3.1536E7, 1.5463008E9}}
\item{percent_diet}{Percent Observed Diet () \link{0.0, 81.9}}
\item{diet_species_cohort}{Diet (Bird species cohort) () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_B_AS_DIET/index.html}
}
\usage{
cciea_B_AS_DIET
}
\description{
No CC Rhinoceros auklet diet: Data from Washington Rhinoceros Auklet Ecology Project; contact tom.good@noaa.gov before citing or distributing these data. Diets of rhinoceros auklet chicks (\% occurrence) calculated from bill loads of returning adults to the colony at Destruction Island, WA. No CC Common murre diet. Data from Hatfield Marine Science Center Seabird Oceanography Lab Yaquina Head Seabird Studies; contact Robert Suryan (rob.suryan@oregonstate.edu) before citing or distributing these data. Diets of common murre chicks (\% occurrence) observed as bill loads of returning adults to colonies at Yaquina Head, OR. Ce CC Rhinoceros auklet diet: Data from Oikonos Ecosystem Knowledge Ano Nuevo Seabird Conservation and Restoration Project; contact Ryan Carle (ryan@oikonos.org) before citing or distributing these data. Diets of rhinoceros auklet chicks (\% occurrence) calculated from bill loads of returning adults to the colony at Ano Nuevo Island, CA. Ce CC: Brandt's cormorant diet. Data from Point Blue Conservation Science collected on Southeast Farallon Island in collaboration with the Farallon Islands National Wildlife Refuge (USFWS); contact Dr. Jaime Jahncke (jjahncke@pointblue.org) before citing or distributing these data. Diet is percent occurrence of fish species in the diets of adult birds that are provisioning chicks calculated from bill loads of adults returning to the colony at Southeast Farallon Island, CA.
}
\concept{dataset_erddap}
\keyword{datasets}
