% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_B_PR_ANOM}
\alias{cciea_B_PR_ANOM}
\title{Seabird Productivity}
\format{
A data frame with 80 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [1.1676096E9, 1.5778368E9]}
\item{productivity_anomaly}{Productivity Anomaly () [-0.969, 1.511]}
\item{species_cohort}{Species cohort () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_B_PR_ANOM/index.html}
}
\usage{
cciea_B_PR_ANOM
}
\description{
Data from Hatfield Marine Science Center Seabird Oceanography Lab Yaquina Head Seabird Studies; contact Robert Suryan (rob.suryan@oregonstate.edu) before citing or distributing these data.
}
\concept{dataset_erddap}
\keyword{datasets}
