% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_EI_CP_SIMP}
\alias{cciea_EI_CP_SIMP}
\title{Coastal pelagics-Simp}
\format{
A data frame with 15 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [8.836128E8, 1.325376E9]}
\item{simpson_diversity}{Simpson Diversity (1-lambda) [0.3012306, 0.4404312]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_CP_SIMP/index.html}
}
\usage{
cciea_EI_CP_SIMP
}
\description{
See: https://www.noaa.gov/iea/Assets/iea/california/Report/pdf/Ecological\%20Integrity\%20Status\%20CCIEA\%202012.pdf
}
\concept{dataset_erddap}
\keyword{datasets}
