% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciea_EI_CP_SPNO.R
\docType{data}
\name{cciea_EI_CP_SPNO}
\alias{cciea_EI_CP_SPNO}
\title{Coastal pelagics-Spp No}
\format{
A data frame with 15 rows and 2 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{8.836128E8, 1.325376E9}}
\item{species_number}{Species Number (No. of species) \link{2.530864, 4.763441}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_CP_SPNO/index.html}
}
\usage{
cciea_EI_CP_SPNO
}
\description{
See: https://www.noaa.gov/iea/Assets/iea/california/Report/pdf/Ecological\%20Integrity\%20Status\%20CCIEA\%202012.pdf
}
\concept{dataset_erddap}
\keyword{datasets}
