% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_EI_CRAB_FINF}
\alias{cciea_EI_CRAB_FINF}
\title{Finfish scavengers:total biomass ratio}
\format{
A data frame with 48 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{1.0413792E9, 1.5147648E9}}
\item{population}{Population () []}
\item{biomass_ratio}{ () \link{0.03598382, 0.3129661}}
\item{Seup}{Confidence Interval, Upper () \link{0.04058898, 0.4040344}}
\item{Selo}{Confidence Interval, Lower () \link{0.02916655, 0.2584166}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_CRAB_FINF/index.html}
}
\usage{
cciea_EI_CRAB_FINF
}
\description{
Source Data: Groundfish data are from the NMFS U.S. West Coast Groundfish Bottom Trawl Survey (https://www.nwfsc.noaa.gov/research/divisions/fram/groundfish/bottom_trawl.cfm) and were downloaded from the FRAM data warehouse at https://www.nwfsc.noaa.gov/data/map with the following API query:  https://www.nwfsc.noaa.gov/data/api/v1/source/trawl.catch_fact/selection.csv?variables=program,trawl_id,date_dim$year,date_dim$yyyymmdd,vessel,performance,year_stn_invalid,depth_m,latitude_dd,longitude_dd,scientific_name,common_name,species_category,partition,total_catch_numbers,total_catch_wt_kg,cpue_kg_per_ha_der,cpue_numbers_per_ha_der,"actual_station_design_dim$station_invalid_survey_year", "actual_station_design_dim$reason_station_invalid","reason_stn_invalid (target station)", "year_stn_invalid" Additional Calculations:  Data are area-weighted average crab:finfish biomass ratios from the southern US border to the northern US border. For details, see Tolimieri et al. 2014 (https://www.noaa.gov/iea/CCIEA-Report/pdf/index.html)
}
\concept{dataset_erddap}
\keyword{datasets}
