% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_EI_DOMACID_MON_crab}
\alias{cciea_EI_DOMACID_MON_crab}
\title{Monthly Maximum Domoic acid levels in dungeness crab, West Coast}
\format{
A data frame with 197 rows and 3 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{6.323616E8, 1.573776E9}}
\item{region}{Region () []}
\item{domoic_acid_concentration}{Maximum Domoic Acid Concentration (ppm) \link{0.0, 270.0}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_DOMACID_MON_crab/index.html}
}
\usage{
cciea_EI_DOMACID_MON_crab
}
\description{
Data on domoic acid concentrations in razor clams are from Ms. Audrey Coyne (audrey.coyne@doh.wa.gov; Washington State Department of Health); these data are compiled from tests conducted by a variety of Tribal, State, and County partners on Washington beaches. Domoic acid levels are reported in ppm. Sample testing frequency is irregular as it depends on the timing of proposed recreational razor clamming digs by Washington State Department of Fish and Wildlife and prevalence of recent detections. We present data as maximum monthly values compiled from samples in a variety of locations on the Washington State outer coast beaches.
}
\concept{dataset_erddap}
\keyword{datasets}
