% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_EI_JELLY_OW}
\alias{cciea_EI_JELLY_OW}
\title{Jellyfish Biomass, Oregon and Washington}
\format{
A data frame with 72 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [9.151488E8, 1.4516064E9]}
\item{species_group}{Species Group () []}
\item{mean_cpue}{Mean CPUE (ln(catch+1)) [0.48, 10919.47]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_JELLY_OW/index.html}
}
\usage{
cciea_EI_JELLY_OW
}
\description{
Source Data: Dr. Ric Brodeur (NOAA; rick.brodeur@noaa.gov), derived from surface trawls conducted as part of the BPA Plume Survey. Additional calculations by C. Barcelo (OSU). Additional Calculations: Geometric mean catch per unit effort (CPUE, #/sq. km) from Newport, OR (44.6 deg N, 124.0 deg W) to Father and Son, WA (48.24 deg. N, 124.7 deg W) by year; Numbers of individuals were recorded for each species caught in each haul and were standardized by the horizontal distance sampled by the towed net as CPUE. Yearly abundance data were obtained by combining (summing) the standardized count data of each species captured during June for each year. All CPUE estimates were log transformed.
}
\concept{dataset_erddap}
\keyword{datasets}
