% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_EI_RREAS_diversity_grid}
\alias{cciea_EI_RREAS_diversity_grid}
\title{Rockfish Recruitment and Ecosystem Assessement Survey, Biodiversity: Grid format}
\format{
A data frame with 783 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [6.31152E8, 1.5147648E9]}
\item{taxa}{Taxa () []}
\item{abundance}{Total Abundance (ln(CPUE + 1)) [1.390595, 12.2398]}
\item{abundance_error}{Abundance Error (ln(CPUE + 1)) [1.744331, 13.34593]}
\item{richness}{Species Richness (number of species) [1.5, 21.98214]}
\item{richness_error}{Species Richness Error (number of species) [1.453661, 7.003501]}
\item{diversity}{Shannon-Weaver Diversity () [0.0696443, 1.690858]}
\item{diversity_error}{Shannon-Weaver Diversity Error () [0.1325587, 0.8078451]}
\item{evenness}{Pielou's Evenness () [0.03166048, 0.7764255]}
\item{evenness_error}{Pielou's Evenness Error () [0.05746754, 0.4711267]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_RREAS_diversity_grid/index.html}
}
\usage{
cciea_EI_RREAS_diversity_grid
}
\description{
Biodiversity indices (total abundance, species richness, Shannon-Weaver diversity, Pielou's evenness) of all taxa (also young-of-the-year rockfish and groundfish taxa, and forage taxa) caught in the RREAS (Rockfish Recruitment and Ecosystem Assessment Survey) mid-water trawl. This is a regional average of all net haul stations located from Point Reyes to Monterey CA.
}
\concept{dataset_erddap}
\keyword{datasets}
