% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_EI_WH_ENT}
\alias{cciea_EI_WH_ENT}
\title{Whale Entanglements}
\format{
A data frame with 420 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [9.466848E8, 1.5778368E9]}
\item{species}{Species () []}
\item{number_of_entanglements}{Number of Entanglements () [0.0, 55.0]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_EI_WH_ENT/index.html}
}
\usage{
cciea_EI_WH_ENT
}
\description{
Whale Entanglements
}
\concept{dataset_erddap}
\keyword{datasets}
