% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_GF_ABND}
\alias{cciea_GF_ABND}
\title{Groundfish Abundance Index}
\format{
A data frame with 130 rows and 3 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [1.0413792E9, 1.4200704E9]}
\item{species_group}{Species Group () []}
\item{abundance_index}{ (abundance index) [297.1226, 138590.0]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_GF_ABND/index.html}
}
\usage{
cciea_GF_ABND
}
\description{
Source Data: Groundfish data are from the NMFS U.S. West Coast Groundfish Bottom Trawl Survey (https://www.nwfsc.noaa.gov/research/divisions/fram/groundfish/bottom_trawl.cfm) and were provided by Dr. Todd Hay and Ms. Beth Horness (NOAA). Additional Calculations: Reference Cope and Haltuch (2014) CCIEA PHASE III REPORT 2013: ECOSYSTEM COMPONENTS - GROUNDFISH report for full explanation of what sources of information were used for each species.
}
\concept{dataset_erddap}
\keyword{datasets}
