% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_GF_STAT}
\alias{cciea_GF_STAT}
\title{Groundfish Stock Status}
\format{
A data frame with 7714 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [-2.9663712E9, 1.4200704E9]}
\item{species_group}{Species Group () []}
\item{relative_stock_status}{ (relative biomass) [0.0498108, 1.25279]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_GF_STAT/index.html}
}
\usage{
cciea_GF_STAT
}
\description{
Source Data: Dr. Jason Cope (NOAA; jason.cope@noaa.gov), derived from NMFS stock assessments (https://www.nwfsc.noaa.gov/research/divisions/fram/popeco/assessment.cfm). Additional Calculations: Reference Cope and Haltuch (2014) CCIEA PHASE III REPORT 2013: ECOSYSTEM COMPONENTS - GROUNDFISH report for full explanation of what sources of information were used for each species.
}
\concept{dataset_erddap}
\keyword{datasets}
