% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_HB_FLO1_CH}
\alias{cciea_HB_FLO1_CH}
\title{Streamflow - Annual 1-day maximum flow anomaly, Chinook ESU's}
\format{
A data frame with 2560 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [3.471552E8, 1.5778368E9]}
\item{flow_anomaly_1_day_max}{Annual 1-day maximum flow anomaly (Annual Anomaly) [-1.902059, 3.428976]}
\item{Seup}{Confidence Interval, Upper (Annual Anomaly) [-1.681833, 3.731741]}
\item{Selo}{Confidence Interval, Lower (Annual Anomaly) [-2.152806, 3.126341]}
\item{location}{Location () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_HB_FLO1_CH/index.html}
}
\usage{
cciea_HB_FLO1_CH
}
\description{
Streamflow is measured using active USGS gages (https://waterdata.usgs.gov/nwis/sw) with records that meet or exceed 30 years in duration.Average daily values from 213 gages were used to calculate annual 1-day maximum flows.  These indicators correspond to flow parameters to which salmon populations are most sensitive.  Standardized anomalies of time series from individual gages were then averaged to obtain weighted averages for ecoregions (for which HUC-8 area served as a weighting factor) and for the entire California current (weighted by ecoregion area).
}
\concept{dataset_erddap}
\keyword{datasets}
