% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_HB_FLO7}
\alias{cciea_HB_FLO7}
\title{Streamflow - Annual 7-day minimum flow anomaly}
\format{
A data frame with 960 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [3.471552E8, 1.5778368E9]}
\item{flow_anomaly_7_day_min}{Annual 7-day minimum flow anomaly (Annual Anomaly) [-1.140393, 1.916366]}
\item{Seup}{95\% credible interval upper bound (Annual Anomaly) [-0.7810067, 2.257912]}
\item{Selo}{95\% credible interval lower bound (Annual Anomaly) [-1.502816, 1.575647]}
\item{location}{Location () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_HB_FLO7/index.html}
}
\usage{
cciea_HB_FLO7
}
\description{
Streamflow is measured using active USGS gages (https://waterdata.usgs.gov/nwis/sw) with records that meet or exceed 30 years in duration.Average daily values from 213 gages were used to calculate annual 7-day minimum flows.  These indicators correspond to flow parameters to which salmon populations are most sensitive.  Standardized anomalies of time series from individual gages were then averaged to obtain weighted averages for ecoregions (for which HUC-8 area served as a weighting factor) and for the entire California current (weighted by ecoregion area).
}
\concept{dataset_erddap}
\keyword{datasets}
