% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_HB_SWE}
\alias{cciea_HB_SWE}
\title{Snow Water Equivalent}
\format{
A data frame with 1332 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [-1.6094592E9, 1.5778368E9]}
\item{SWE_anomaly}{1 Apr SWE Anomaly (Annual Anomaly) [-2.187893, 2.455789]}
\item{Seup}{95\% credible interval upper bound (Annual Anomaly) [-1.721963, 3.500461]}
\item{Selo}{95\% credible interval lower bound (Annual Anomaly) [-2.957523, 2.095322]}
\item{location}{Location () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_HB_SWE/index.html}
}
\usage{
cciea_HB_SWE
}
\description{
Snow-water equivalent data were derived from the California Department of Water Resources snow survey (https://cdec.water.ca.gov/) and the Natural Resources Conservation Service's SNOTEL sites in WA, OR, CA and ID (https://www.wcc.nrcs.usda.gov/snow/). Anomalies of April 1 snow-water equivalents (SWE) for the CCE, calculated as an area-weighted average of data from 5 ecoregions. SWE is a measure of the total water available in snowpack. Measurements on April 1st are considered the best indicator of maximum extent of snowpack.
}
\concept{dataset_erddap}
\keyword{datasets}
