% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciea_HD_ESI_VESS.R
\docType{data}
\name{cciea_HD_ESI_VESS}
\alias{cciea_HD_ESI_VESS}
\title{Annual Effective Shannon Index (ESI) for various fleet groupings and West Coast Ports}
\format{
A data frame with 1330 rows and 3 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{3.471552E8, 1.5147648E9}}
\item{average_ESI}{Average ESI (Average ESI) \link{1.23, 9.44}}
\item{vessel_category}{Vessel category () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_HD_ESI_VESS/index.html}
}
\usage{
cciea_HD_ESI_VESS
}
\description{
Indices for annual average Effective Shannon Index (ESI) values for various fleet groupings and West Coast Ports Data derived from Pacific Fisheries Information Network (PacFIN; https://pacfin.psmfc.org/) and Alaska Fisheries Information Network (AKFIN; https://www.akfin.org/). Aggregation and manipulaton of vessel level annual revenue data to create effective shannon index. For details, see: Kasperski, S. and D.S. Holland 2013. Income Diversification and Risk for Fishermen. Proceedings of the National Academy of Science. 100(6):2076-2081.  doi: 10.1073/pnas.1212278110; Holland, D.S. and S. Kasperski 2016. The Impact of Access Restrictions on Fishery Income Diversification of US West Coast Fishermen. Forthcoming in Coastal Management.
}
\concept{dataset_erddap}
\keyword{datasets}
