% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_HD_Fish_Rel}
\alias{cciea_HD_Fish_Rel}
\title{Commercial Fishing Reliance Index}
\format{
A data frame with 598 rows and 4 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [1.1045376E9, 1.4832288E9]}
\item{fishing_reliance}{Commercial fishing reliance index (Index) [-0.941, 29.54986]}
\item{location}{Location () []}
\item{region}{Region () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_HD_Fish_Rel/index.html}
}
\usage{
cciea_HD_Fish_Rel
}
\description{
Commercial fishing  data were provided by Dr. Karma Norman (NOAA), and are derived from state reported fish ticket data as maintained by the Pacific Fishery Information Network (PacFIN) (https://pacfin.psmfc.org/).
}
\concept{dataset_erddap}
\keyword{datasets}
