% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciea_HD_Soc_Vuln.R
\docType{data}
\name{cciea_HD_Soc_Vuln}
\alias{cciea_HD_Soc_Vuln}
\title{Coastal community social vulnerability index}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{9.466848E8, 1.4200704E9}}
\item{social_vulnerability}{Social Vulnerability Index () \link{-6.23433, 13.66654}}
\item{location}{Location () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_HD_Soc_Vuln/index.html}
}
\usage{
cciea_HD_Soc_Vuln
}
\description{
Source Data: Community social vulnerability index (CSVI) data were provided by Dr. Karma Norman (NOAA), and are derived from the American Community Survey (ACS) associated with the US Census (https://factfinder.census.gov/faces/nav/jsf/pages/index.xhtml).  Although decennial census data represent an estimate at a specific date, the ACS data are period estimates, collected over an entire year and are potentially averaged over varying time periods, depending on the size of the geographic area (U.S. Census Bureau, 2009).  Given that communities of interest include geographic areas with populations less than 20,000, ACS data at the Census-Designate Place level or for place-based communities, are averaged over a five-year period., Additional Calculations:  The Community Social Vulnerability Index (CSVI) is derived from social vulnerability indices (e.g., personal disruption, poverty, population composition, housing characteristics, housing disruption, labor force structure, and natural resource labor force).
}
\concept{dataset_erddap}
\keyword{datasets}
