% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_ARG}
\alias{cciea_OC_ARG}
\title{Aragonite Saturation}
\format{
A data frame with 552 rows and 3 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [8.836128E8, 1.6067808E9]}
\item{location}{Location () []}
\item{aragonite_saturation}{ (relative to 1) [0.4761976, 2.132431]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_ARG/index.html}
}
\usage{
cciea_OC_ARG
}
\description{
Source Data: Data are derived from methods in Juranek et al. 2009: Juranek, L.W., Feely, R.A., Peterson, W.T., Alin, S.R., Hales, B., Lee, K., Sabine, C.L. and Peterson, J., 2009. A novel method for determination of aragonite saturation state on the continental shelf of central Oregon using multi-parameter relationships with hydrographic data. Geophysical Research Letters, 36(24). Additional Calculations: Data are derived from methods in Juranek et al. 2009: Juranek, L.W., Feely, R.A., Peterson, W.T., Alin, S.R., Hales, B., Lee, K., Sabine, C.L. and Peterson, J., 2009. A novel method for determination of aragonite saturation state on the continental shelf of central Oregon using multi-parameter relationships with hydrographic data. Geophysical Research Letters, 36(24).
}
\concept{dataset_erddap}
\keyword{datasets}
