% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_BEUTI}
\alias{cciea_OC_BEUTI}
\title{Monthly Biologically Effective Upwelling Transport Index (BEUTI)}
\format{
A data frame with 2406 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [5.692032E8, 1.6210368E9]}
\item{beuti}{Biologically Effective Upwelling Transport Index (mmol s^-1 m^-1) [-22.082, 57.187]}
\item{latitude}{Latitude (degrees_north) [33, 45]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_BEUTI/index.html}
}
\usage{
cciea_OC_BEUTI
}
\description{
BEUTI provides estimates of vertical nitrate flux near the coast (i.e., the amount of nitrate upwelled/downwelled), which may be more relevant than upwelling strength when considering some biological responses.
}
\concept{dataset_erddap}
\keyword{datasets}
