% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_CBAN}
\alias{cciea_OC_CBAN}
\title{Northern copepod biomass anomaly 44.6N}
\format{
A data frame with 213 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [8.309088E8, 1.4148E9]}
\item{biomass_anomaly}{Biomass Anomaly () [-1.541188, 1.013472]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_CBAN/index.html}
}
\usage{
cciea_OC_CBAN
}
\description{
See: https://www.noaa.gov/iea/Assets/iea/california/Report/pdf/Ecological\%20Integrity\%20Status\%20CCIEA\%202012.pdf
}
\concept{dataset_erddap}
\keyword{datasets}
