% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_CUTI}
\alias{cciea_OC_CUTI}
\title{Monthly Coastal Upwelling Transport Index (CUTI)}
\format{
A data frame with 1191 rows and 3 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [5.679936E8, 1.6094592E9]}
\item{cuti}{Coastal Upwelling Transport Index (m^2 s^(-1)) [-1.252, 2.751]}
\item{latitude}{Latitude (degrees_north) [33, 45]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_CUTI/index.html}
}
\usage{
cciea_OC_CUTI
}
\description{
CUTI provides estimates of vertical transport near the coast (i.e., upwelling/downwelling). It was developed as a more accurate alternative to the previously available Bakun Index.
}
\concept{dataset_erddap}
\keyword{datasets}
