% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_DO}
\alias{cciea_OC_DO}
\title{Dissolved Oxygen}
\format{
A data frame with 5401 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [-6.31152E8, 1.6067808E9]}
\item{location}{Location () []}
\item{dissolved_oxygen}{ (ml/L) [0.62, 6.66876]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_DO/index.html}
}
\usage{
cciea_OC_DO
}
\description{
Source Data: Dr. Bill Peterson, NOAA (bill.peterson@noaa.gov); https://www.nwfsc.noaa.gov/research/divisions/fe/estuarine/oeip/index.cfm; Source Data: Isaac Schroeder (NOAA; isaac.schroeder@noaa.gov); derived from CalCOFI surveys (http://calcofi.org/)
}
\concept{dataset_erddap}
\keyword{datasets}
