% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_LUSI}
\alias{cciea_OC_LUSI}
\title{Length of Upwelling Season Index (LUSI)}
\format{
A data frame with 324 rows and 3 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [-9.46944E7, 1.5778368E9]}
\item{lusi}{Length of Upwelling Season Index (days) [58.0, 365.0]}
\item{latitude}{Latitude (degrees_north) [33, 48]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_LUSI/index.html}
}
\usage{
cciea_OC_LUSI
}
\description{
Length of Upwelling Season Index
}
\concept{dataset_erddap}
\keyword{datasets}
