% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_MHW_EV}
\alias{cciea_OC_MHW_EV}
\title{California Current Marine Heatwave Event List}
\format{
A data frame with 413 rows and 8 variables:
\describe{
\item{time}{Start time (seconds since 1970-01-01T00:00:00Z) \link{3.874176E8, 1.6003872E9}}
\item{blob_id}{Blob Id () []}
\item{max_area}{Maximum Area (km^2) \link{402937.0, 9759080.0}}
\item{duration}{Duration (days) \link{0, 757}}
\item{max_intensity}{Maximum Intensity (STDEVs from normal) \link{1.43471, 2.56513}}
\item{mean_intensity}{Mean Intensity (STDEVs from normal) \link{1.43471, 2.41716}}
\item{min_dist_to_coast}{Minimum distance to coast (km) \link{5.95226, 2772.68}}
\item{max_index}{Maximum MHW index () \link{0.0, 0.771722}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_MHW_EV/index.html}
}
\usage{
cciea_OC_MHW_EV
}
\description{
Marine heatwaves, or MHWs, occur when ocean temperatures are much warmer than usual for an extended period of time; they are specifically defined by differences in expected temperatures for the location and time of year. MHWs are a growing field of study worldwide because of their effects on ecosystem structure, biodiversity, and regional economies. Developed by oceanographers from NOAA Fisheries' Southwest Fisheries Science Center as an experimental tool for natural resource managers, the California Current MHW Tracker is a program designed to understand, describe, and provide a historical context for the 2014-16 blob. It also produces a range of indices that could help forecast or predict future MHWs expected to impact our coast.
}
\concept{dataset_erddap}
\keyword{datasets}
