% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_MHW_regions}
\alias{cciea_OC_MHW_regions}
\title{California Current Marine Heatwave, EEZ regions Daily}
\format{
A data frame with 345316 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [3.786912E8, 1.609416E9]}
\item{heatwave_cover}{Heatwave cover (\%) [0.0, 100.0]}
\item{intensity}{Intensity (degree_C) [1.290018, 3.802803]}
\item{distance}{Distance (km) [0.5110544, 3201.752]}
\item{avedegday}{Cumulative Intensity (degree_C) [0.0, 1087.539]}
\item{avedegdayclim}{Average Day Climatology (degree_C) [35.65968, 138.4911]}
\item{region}{Region () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_MHW_regions/index.html}
}
\usage{
cciea_OC_MHW_regions
}
\description{
Marine heatwaves, or MHWs, occur when ocean temperatures are much warmer than usual for an extended period of time; they are specifically defined by differences in expected temperatures for the location and time of year. MHWs are a growing field of study worldwide because of their effects on ecosystem structure, biodiversity, and regional economies. Developed by oceanographers from NOAA Fisheries' Southwest Fisheries Science Center as an experimental tool for natural resource managers, the California Current MHW Tracker is a program designed to understand, describe, and provide a historical context for the 2014-16 blob. It also produces a range of indices that could help forecast or predict future MHWs expected to impact our coast.
}
\concept{dataset_erddap}
\keyword{datasets}
