% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_NOI}
\alias{cciea_OC_NOI}
\title{Extratropical-based Northern Oscillation Index}
\format{
A data frame with 877 rows and 4 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{-6.929712E8, 1.6107984E9}}
\item{NOI}{Extratropical-based Northern Oscillation Index (Index) \link{-12.164, 8.678}}
\item{SOIX}{Extratropical-bassed Southern Oscillation Index (Index) \link{-8.9, 8.661}}
\item{SOI}{Southern Oscillation Inde (Index) \link{-6.898, 5.355}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_NOI/index.html}
}
\usage{
cciea_OC_NOI
}
\description{
Northern Oscillation Index - The NOI (extratropical-based Northern Oscillation Index is an index of midlatitude climate fluctuations that shows interesting relationships with fluctuations in marine ecosystems and populations. It reflects the variability in equatorial and extratropical teleconnections and represent a wide range of local and remote climate signals.
}
\concept{dataset_erddap}
\keyword{datasets}
