% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_NPH_JF}
\alias{cciea_OC_NPH_JF}
\title{North Pacific High Area, January - February Mean}
\format{
A data frame with 55 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [-9.46944E7, 1.6094592E9]}
\item{nph_area}{North Pacific High Area, Jan-Feb Mean (10^6 km^2) [0.00561367, 5.806523]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_NPH_JF/index.html}
}
\usage{
cciea_OC_NPH_JF
}
\description{
Variations in large-scale atmospheric forcing influence upwelling dynamics and ecosystem productivity in the California Current System.The area of the North Pacific High (NPH) is characterized by areal extent of the 1020 Pa isobar. Winter values (January - February mean) of the NPH area can be used as an ecosystem pre-conditioning index (Schroeder et al., doi:10.1002/grl.50100).The area of the NPH are calculated from monthly mean sea level pressure (SLP) fields created by the U.S. Navy Fleet Numerical Meterology and Oceanography Center (FNMOC). Monthly SLP data available at https://upwell.pfeg.noaa.gov/erddap/index.html , search for Dataset ID: erdlasFnWPr. The area is the areal extent of the 1020 hPa contour for a given month. The January - February mean is the average of the January and February areas for a given year. The NRT ROMS temperature data downloaded from the UCSC website (http://oceanmodeling.pmc.ucsc.edu:8080/thredds) has grid points with 2 m temperature data reported as Not-A-Number, these values are  2 m temperatu. Tdata at these grid points are obtained by extrapolating the data to 2 m depth using the interp1d routine from the Python library scipy.interpolate sub-package.
}
\concept{dataset_erddap}
\keyword{datasets}
