% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciea_OC_SL1.R
\docType{data}
\name{cciea_OC_SL1}
\alias{cciea_OC_SL1}
\title{Monthly Sea Level: South Beach OR (44.6N 124.0W)}
\format{
A data frame with 588 rows and 7 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{-9.34848E7, 1.4501376E9}}
\item{sea_level}{Sea Level (mm) \link{2522.033, 3236.0}}
\item{std_error}{Standard Error (mm) \link{0.0, 0.0}}
\item{station}{ () []}
\item{latitude}{Latitude (degrees_north) \link{44.6, 44.6}}
\item{longitude}{Longitude (degrees_east) \link{124.0, 124.0}}
\item{depth}{Depth (m) []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_SL1/index.html}
}
\usage{
cciea_OC_SL1
}
\description{
The California Current Large Marine Ecosystem (CCLME) is primarily driven by bottom-up physical oceanographic processes, thus understanding trends in the physical state can inform our knowledge of ecosystem processes and management of ecosystem services.
}
\concept{dataset_erddap}
\keyword{datasets}
