% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_SL2}
\alias{cciea_OC_SL2}
\title{Monthly Sea Level: San Francisco CA (37.8N 122.5W)}
\format{
A data frame with 1392 rows and 7 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{-2.2077792E9, 1.4501376E9}}
\item{sea_level}{Sea Level (mm) \link{2466.067, 3081.964}}
\item{std_error}{Standard Error (mm) \link{0.0, 0.0}}
\item{station}{ () []}
\item{latitude}{Latitude (degrees_north) \link{37.8, 37.8}}
\item{longitude}{Longitude (degrees_east) \link{122.5, 122.5}}
\item{depth}{Depth (m) []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_SL2/index.html}
}
\usage{
cciea_OC_SL2
}
\description{
The California Current Large Marine Ecosystem (CCLME) is primarily driven by bottom-up physical oceanographic processes, thus understanding trends in the physical state can inform our knowledge of ecosystem processes and management of ecosystem services.
}
\concept{dataset_erddap}
\keyword{datasets}
