% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_SST3}
\alias{cciea_OC_SST3}
\title{SST Buoy 46025 (33.7N 119.1W)}
\format{
A data frame with 445 rows and 2 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{3.864672E8, 1.6121376E9}}
\item{SST}{Sea Surface Temperature (degree_C) \link{12.83, 22.44}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_SST3/index.html}
}
\usage{
cciea_OC_SST3
}
\description{
The National Data Buoy Center (NDBC) distributes meteorological data from moored buoys maintained by NDBC and others. The data is from NOAA NDBC. It has been reformatted by NOAA Coastwatch, West Coast Node and then monthly averaged. This dataset has both historical data and near real time data.
}
\concept{dataset_erddap}
\keyword{datasets}
