% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_STI}
\alias{cciea_OC_STI}
\title{Spring Transition Index (STI)}
\format{
A data frame with 648 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [-9.46944E7, 1.5778368E9]}
\item{sti}{Spring Transition Index (yearday) [1.0, 174.0]}
\item{latitude}{Latitude (degrees_north) [33, 48]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_STI/index.html}
}
\usage{
cciea_OC_STI
}
\description{
Spring Transition Index
}
\concept{dataset_erddap}
\keyword{datasets}
