% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciea_OC_TUMI.R
\docType{data}
\name{cciea_OC_TUMI}
\alias{cciea_OC_TUMI}
\title{Total Upwelling Magnitude Index (TUMI)}
\format{
A data frame with 324 rows and 3 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{-9.46944E7, 1.5778368E9}}
\item{tumi}{Total Upwelling Magnitude Index (m^3/s/100m coastline) \link{564.6994, 55104.24}}
\item{latitude}{Latitude (degrees_north) \link{33, 48}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_TUMI/index.html}
}
\usage{
cciea_OC_TUMI
}
\description{
Total Upwelling Magnitude Index
}
\concept{dataset_erddap}
\keyword{datasets}
