% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_UI3}
\alias{cciea_OC_UI3}
\title{Upwelling Index, 33N 119W, monthly}
\format{
A data frame with 649 rows and 4 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [-9.46944E7, 1.6094592E9]}
\item{upwelling_index}{Upwelling Index, 33N 119W, monthly (m^3/s/100m coastline) [-39.515, 292.556]}
\item{nobs}{Number of oberservations in month () [87, 124]}
\item{stdev}{Standard deviation (m^3/s/100m coastline) [23.955, 337.93]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_UI3/index.html}
}
\usage{
cciea_OC_UI3
}
\description{
Upwelling index computed from 1-degree FNMOC sea level pressure at 33 degrees of latitude. The coastal Upwelling Index is an index of the strength of the wind forcing on the ocean which has been used in many studies of the effects of ocean variability on the reproductive and recruitment success of many fish and invertebrate species.
}
\concept{dataset_erddap}
\keyword{datasets}
