% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_VWND3}
\alias{cciea_OC_VWND3}
\title{Meridional Wind Buoy 46025 (33.7N 119.1W)}
\format{
A data frame with 449 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [3.864672E8, 1.6198272E9]}
\item{vwnd}{Wind Speed, Meridional (m s-1) [-2.906052, 0.8122222]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_VWND3/index.html}
}
\usage{
cciea_OC_VWND3
}
\description{
The National Data Buoy Center (NDBC) distributes meteorological data from moored buoys maintained by NDBC and others. The data is from NOAA NDBC. It has been reformatted by NOAA Coastwatch, West Coast Node and then monthly averaged. This dataset has both historical data and near real time data.
}
\concept{dataset_erddap}
\keyword{datasets}
