% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_SM_ORWA_CO_CND}
\alias{cciea_SM_ORWA_CO_CND}
\title{Coho Condition, Oregon/Washington}
\format{
A data frame with 65 rows and 4 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{1.262304E8, 1.3885344E9}}
\item{population}{Population/Season () []}
\item{cond_pct_nat}{Proportion of natural fish to total fish returning () \link{0.9189854, 99.35072}}
\item{cond_pop_gr}{Proportional change in abundance between cohorts () \link{0.4980789, 2.20522}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_SM_ORWA_CO_CND/index.html}
}
\usage{
cciea_SM_ORWA_CO_CND
}
\description{
Source Data:  Various; see Wells et al. 2014, Table S6. For Oregon and Washington ESUs, data were obtained from the NWFSC's "Salmon Population Summary" database (https://www.webapps.nwfsc.noaa.gov/sps), with additional data for Oregon Coast coho salmon (Oregon Department of Fish and Wildlife, https://oregonstate.edu/dept/ODFW/spawn/data.htm), and from PFMC (2012) for the Upper Columbia Summer/Fall-run Chinook Salmon. Additional Calculations: Proportion of natural-origin spawners, computed for a single population as the fraction NN/NT, where NN is the number of naturally-origin spawners, and NT is the total number of spawners. Population fractions were then averaged across the populations within the ESU, weighted by total spawner abundance.
}
\concept{dataset_erddap}
\keyword{datasets}
