% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeTimeSeries.R
\name{GetNorm}
\alias{GetNorm}
\title{Normalize Time Series}
\usage{
GetNorm(X, Y, Name)
}
\arguments{
\item{X}{time as year, year-month or year-month-day}

\item{Y}{value to plot over time}

\item{Name}{title of plot}
}
\value{

}
\description{
Normalize Time Series
}
