% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundfish.R
\docType{data}
\name{groundfish}
\alias{groundfish}
\title{Groundfish diversity}
\format{
Format with columns...

\itemize{
\item year: ...
\item index: ...
\item SEup: ...
\item SElo: ...
\item timeseries: ...
\item metric: ...
}
}
\usage{
groundfish
}
\description{
Description...
}
\details{
Example for more info: More information available at \url{https://noaa-edab.github.io/tech-doc/aquaculture.html}
}
\keyword{datasets}
