% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{newportCTD}
\alias{newportCTD}
\title{Newport Hydrographic Line CTD Data, NH05 and NH25}
\format{
A data frame with 1002090 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [8.589024E8, 1.5099264E9]}
\item{station}{Station () []}
\item{depth}{Depth (m) [0, 743]}
\item{temperature}{Water Temperature (degrees C) [5.7691, 18.2575]}
\item{salinity}{Salinity (PSU) [23.8151, 34.7945]}
\item{density}{Density () [17.1925, 27.1704]}
\item{oxygen}{Oxygen () [0.2496, 12.01967]}
\item{latitude}{Latitude (degrees_north) [44.391, 44.70633]}
\item{longitude}{Longitude (degrees_east) [-124.6712, -124.106]}
\item{project}{Project () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/newportCTD/index.html}
}
\usage{
newportCTD
}
\description{
Newport Hydrographic Line CTD Data, NH05 and NH25
}
\concept{dataset_erddap}
\keyword{datasets}
