% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{newportDOArag}
\alias{newportDOArag}
\title{Newport Hydrographic Line Dissolved Oxygen and Aragonite Saturation, NH05 and NH25}
\format{
A data frame with 2656 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [9.019296E8, 1.5094944E9]}
\item{station}{Station () []}
\item{depth}{Depth (m) [40, 150]}
\item{temperature}{Water Temperature (degrees C) [6.832, 14.66645]}
\item{salinity}{Salinity (PSU) [31.8645, 33.9962]}
\item{density}{Density () [24.2092, 26.648]}
\item{oxygen}{Oxygen () [0.54, 6.9]}
\item{dissolved_oxygen}{Dissolved Oxygen (umol/kg) [23.4929, 300.5841]}
\item{aragonite_saturation}{Aragonite Saturation () [0.4761976, 2.132431]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/newportDOArag/index.html}
}
\usage{
newportDOArag
}
\description{
Newport Hydrographic Line Dissolved Oxygen and Aragonite Saturation, NH05 and NH25
}
\concept{dataset_erddap}
\keyword{datasets}
