% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{newportNitrate}
\alias{newportNitrate}
\title{Newport Hydrographic Line Nitrate, NH25}
\format{
A data frame with 498 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [1.0997856E9, 1.4998176E9]}
\item{station}{Station () []}
\item{depth}{Depth (m) [150, 150]}
\item{nitrate}{Nitrate (no3+no2) () [4.5802, 36.2214]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/newportNitrate/index.html}
}
\usage{
newportNitrate
}
\description{
Newport Hydrographic Line Nitrate, NH25
}
\concept{dataset_erddap}
\keyword{datasets}
