% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDomoicAcid.R
\name{Plot_Domoic_Acid_Monthly}
\alias{Plot_Domoic_Acid_Monthly}
\title{Plot domoic acid monthly, original}
\usage{
Plot_Domoic_Acid_Monthly(
  Data.File = NA,
  Date = "Date",
  Location = "Waterbody",
  DA = "DA.ppm",
  Species = "Species",
  spp = NA,
  THRESH = c(30, 20),
  COL = c("black", "darkgrey"),
  Xlim = NA,
  Ylab = "Monthly max (ppm)",
  Ymax = NA,
  Title = NA,
  LegendLoc = "topleft",
  legend.cex = 0.8,
  legend.inset = c(0),
  pt.cex = 0.1,
  LWD = 1
)
}
\arguments{
\item{Data.File}{...}

\item{Date}{...}

\item{Location}{...}

\item{DA}{...}

\item{Species}{...}

\item{spp}{...}

\item{THRESH}{...}

\item{COL}{...}

\item{Xlim}{...}

\item{Ylab}{...}

\item{Ymax}{...}

\item{Title}{...}

\item{LegendLoc}{...}

\item{legend.cex}{...}

\item{legend.inset}{...}

\item{pt.cex}{...}

\item{LWD}{...}
}
\value{

}
\description{
Plot domoic acid monthly, original
}
\concept{plot_original}
