% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series}
\usage{
plot_ts(
  d,
  x = year,
  y = index,
  y_lo = SElo,
  y_hi = SEup,
  x_recent = 5,
  units_recent = "years",
  add_avg = TRUE,
  add_icons = TRUE,
  font_size = 24,
  icon_size = font_size/2,
  color_pts = rgb(225, 118, 44, max = 255),
  color_avg = rgb(63, 173, 213, max = 255),
  color_hilo = "lightgrey",
  alpha_avg = 0.8,
  alpha_hilo = 0.8,
  theme_plot = theme_iea
)
}
\arguments{
\item{d}{Default dataset to use for plotting with
\code{ggplot2::\link[ggplot2]{ggplot}()}. Typically includes columns:
\code{year}, \code{index}, \code{SElo} and \code{SEup}.}

\item{x}{column for x axis from \code{d}, unquoted. Defaults to \code{year}.}

\item{y}{column for y axis from \code{d}, unquoted. Defaults to \code{index}.}

\item{y_lo}{unquoted column for low value to apply to gray ribbon, typically
\code{y - sd(y)}. Defaults to \code{SElo} if present.}

\item{y_hi}{unquoted column for highy value to apply to gray ribbon,
typically \code{y + sd(y)}. Defaults to \code{SEhi} if present.}

\item{x_recent}{duration of \code{x}, typically an integer, for which to consider
as recent period to derive \strong{rectangle} and \strong{icons}. Defaults to \code{5}, as
in years.}

\item{units_recent}{units to describe \code{x}, used in caption.}

\item{add_avg}{whether to add average as a black dashed line and solid lines
indicating the \strong{standard deviation} in the color of \code{color_avg}. Defaults
to \code{TRUE}.}

\item{add_icons}{whether to include icons in the right-hand margin for recent
trend (→, ↗, ↘) and recent average (o, +, -) if stable, increasing, or
decreasing (respectively) relative to the long-term standard deviation.
Defaults to \code{TRUE}.}

\item{font_size}{font size. Defaults to 24.}

\item{icon_size}{size of icon font. Defaults to half the \code{font_size}, 12.}

\item{color_pts}{color for points, as named color (see
\code{\link[grDevices]{colors}()}) or hexadecimal value as provided by red
green blue \code{\link[grDevices]{rgb}()} specification. Defaults to dusty
orange.}

\item{color_avg}{color for standard deviation lines around long-term average
and rectangle between since \code{x_recent}.}

\item{color_hilo}{color for ribbon around \code{y} between values \code{y_lo} and
\code{y_hi}.}

\item{alpha_avg}{transparency (0 to 1) for colored elements of long-term
average: standard deviation and recent rectangle.}

\item{alpha_hilo}{transparency (0 to 1) for ribbon around \code{y} between values
\code{y_lo} and \code{y_hi}.}

\item{theme_plot}{theme function to apply to the ggplot. Defaults to
\code{\link{theme_iea}}.}
}
\value{
This function returns a ggplot object with a caption attribute.
}
\description{
Plot time series in the style of the California Current IEA. Dashed lines are
used in between \strong{missing data} and points colored by \code{color_pts}. Include a
\strong{ribbon} around \code{y} if low (\code{y_lo}) and high (\code{y_hi}) values are provided.
Include \strong{long-term average} if \code{add_avg = TRUE} with average as a black
dashed line and solid lines indicating the \strong{standard deviation} in the
color of \code{color_avg}. Include \strong{rectangle} in the color of \code{color_avg}
between the standard deviation of the long-term average since \code{x_recent} if
provided. Include \strong{icons} in the right-hand margin for recent trend (→, ↗,
↘) and recent average (o, +, -) if stable, increasing, or decreasing
(respectively) relative to the long-term standard deviation.
}
\details{
The attribute \code{caption} includes a textual summary.
}
\examples{
# example time series dataset
ts1

# defaults to include all options
g <- plot_ts(ts1)
g
cat(attr(g, "caption"))

# without SElo or SEhi
g <- plot_ts(ts1[,c("year","index")])
g
cat(attr(g, "caption"))

# without x_recent, add_avg, add_icons
g <- plot_ts(ts1, x_recent=NA, add_icons=F, add_avg=F)
g
cat(attr(g, "caption")) # empty caption without x_recent

}
\concept{plot}
