% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadPlot.R
\name{QuadPlot}
\alias{QuadPlot}
\title{QuadPlot, original}
\usage{
QuadPlot(
  dframe = NULL,
  X = NA,
  Y = NA,
  Name = NA,
  PERIOD = 5,
  d1.col = NULL,
  d2.col = NULL,
  rng = NULL,
  Title = NA,
  plot.type = 1,
  style = 1,
  flip.colors = NULL,
  bg.polys = FALSE,
  Xlab = NULL,
  Ylab = NULL,
  Label = NULL,
  Symbol = c(21, 21, 21, 21, 21, 24, 24, 24, 24, 24, 25, 25, 25, 25, 25),
  bg.color = c("black", "white", "grey", "red", "blue"),
  ol.color = "black",
  Xlim = NA,
  Ylim = NA,
  AxisCex = 0.9,
  LabelCex = 0.9,
  XYLabCex = 0.9,
  LegendCex = 0.8,
  LegendPointCex = 1,
  TitleCex = 1,
  TextCex = 0.8,
  PointCex = 1,
  Period.Label = "Years",
  Bmar = NA,
  Lmar = NA,
  Tmar = NA,
  Rmar = NA,
  AltTxt = NA,
  l.inset = c(0, 0),
  TextOffset = 0.5,
  PlotLegend = NULL
)
}
\arguments{
\item{dframe}{...}

\item{X}{...}

\item{Y}{...}

\item{Name}{...}

\item{PERIOD}{...}

\item{d2.col}{...}

\item{rng}{...}

\item{Title}{...}

\item{plot.type}{...}

\item{style}{...}

\item{flip.colors}{...}

\item{bg.polys}{...}

\item{Xlab}{...}

\item{Ylab}{...}

\item{Label}{...}

\item{Symbol}{...}

\item{bg.color}{...}

\item{ol.color}{...}

\item{Xlim}{...}

\item{Ylim}{...}

\item{AxisCex}{...}

\item{LabelCex}{...}

\item{XYLabCex}{...}

\item{LegendCex}{...}

\item{LegendPointCex}{...}

\item{TitleCex}{...}

\item{TextCex}{...}

\item{PointCex}{...}

\item{Period.Label}{...}

\item{Bmar}{...}

\item{Lmar}{...}

\item{Tmar}{...}

\item{Rmar}{...}

\item{AltTxt}{...}

\item{l.inset}{...}

\item{TextOffset}{...}

\item{PlotLegend}{...}
}
\value{

}
\description{
QuadPlot, original
}
\concept{plot_original}
