% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadplot_means.R
\name{QuadPlot_means}
\alias{QuadPlot_means}
\title{Quadplot means}
\usage{
QuadPlot_means(
  DATA = NULL,
  D1 = NA,
  D2 = NA,
  SE.d1.up = NULL,
  SE.d1.lo = NULL,
  SE.d2.up = NULL,
  SE.d2.lo = NULL,
  SE.d1 = NULL,
  SE.d2 = NULL,
  d1.col = NULL,
  d2.col = NULL,
  rng = NULL,
  Xlim = NA,
  Ylim = NA,
  Title = NA,
  plot.type = 1,
  style = 1,
  flip.colors = NULL,
  bg.polys = FALSE,
  Xlab = NULL,
  Ylab = NULL,
  Label = NA,
  Symbol = c(21, 21, 21, 21, 21, 24, 24, 24, 24, 24, 25, 25, 25, 25, 25),
  bg.color = c("black", "white", "grey", "red", "blue"),
  ol.color = "black",
  AxisCex = 0.9,
  LabelCex = 0.9,
  XYLabCex = 0.9,
  LegendCex = 0.8,
  LegendPointCex = 1,
  TitleCex = 1,
  TextCex = 0.8,
  PointCex = 1,
  error.lwd = NA,
  PERIOD = 5,
  Period.Label = "Years",
  Bmar = NA,
  Lmar = NA,
  Tmar = NA,
  Rmar = NA,
  AltTxt = NA,
  TextOffset = 0.5,
  PlotLegend = NULL
)
}
\arguments{
\item{DATA}{...}

\item{D1}{...}

\item{D2}{...}

\item{SE.d1.up}{...}

\item{SE.d1.lo}{...}

\item{SE.d2.up}{...}

\item{SE.d2.lo}{...}

\item{SE.d1}{...}

\item{SE.d2}{...}

\item{d1.col}{...}

\item{d2.col}{...}

\item{rng}{...}

\item{Xlim}{...}

\item{Ylim}{...}

\item{Title}{...}

\item{plot.type}{...}

\item{style}{...}

\item{flip.colors}{...}

\item{bg.polys}{...}

\item{Xlab}{...}

\item{Ylab}{...}

\item{Label}{...}

\item{Symbol}{...}

\item{bg.color}{...}

\item{ol.color}{...}

\item{AxisCex}{...}

\item{LabelCex}{...}

\item{XYLabCex}{...}

\item{LegendCex}{...}

\item{LegendPointCex}{...}

\item{TitleCex}{...}

\item{TextCex}{...}

\item{PointCex}{...}

\item{error.lwd}{...}

\item{PERIOD}{...}

\item{Period.Label}{...}

\item{Bmar}{...}

\item{Lmar}{...}

\item{Tmar}{...}

\item{Rmar}{...}

\item{AltTxt}{...}

\item{TextOffset}{...}

\item{PlotLegend}{...}
}
\value{

}
\description{
Makes a quadplot but does NOT standardize data first. Requires that you use NormalizeTimeSeries separately.
}
\details{
x = short term trend
y = mean over the evaluation period
the data come out of the normalization in this format
style 1 will graph them as x,y
style 2 will swap the axes internally to y, x.
this is done to maintain some formatting between some different codes
CAN plot standard errors but not clear what they should be at present.
no alt text for standard errors.
PlotLegend can be used to turn the Legend on or Off.  (TRUE or FALSE)
}
