% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_example.R
\docType{data}
\name{ts1}
\alias{ts1}
\title{TimeSeries1 example dataset}
\format{
Format with columns...

\itemize{
\item year: integer year \link{1985 to 2015}
\item index: main value (double)
\item SEup: standard error, high
\item SElo: standard error, low
\item timeseries: "(a) Trend and recent mean"
\item metric: "current.data"
}
}
\usage{
ts1
}
\description{
Example of time series data to be used by \code{\link{plot_ts}}
}
\details{
Example for more info: More information available at \url{https://noaa-edab.github.io/tech-doc/aquaculture.html}
}
\concept{dataset_other}
\keyword{datasets}
