% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciea_HD_Fish_Eng.R
\docType{data}
\name{cciea_HD_Fish_Eng}
\alias{cciea_HD_Fish_Eng}
\title{Commercial Fishing Engagement Index}
\format{
A data frame with 455 rows and 4 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{1.1045376E9, 1.4832288E9}}
\item{fishing_engagement}{Commercial fishing engagement index (Index) \link{-0.19823, 16.83776}}
\item{location}{Location () []}
\item{region}{Region () []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_HD_Fish_Eng/index.html}
}
\usage{
cciea_HD_Fish_Eng
}
\description{
Commercial fishing engagement data were provided by Dr. Karma Norman (NOAA), and are derived from state reported fish ticket data as maintained by the Pacific Fishery Information Network (PacFIN) (https://pacfin.psmfc.org/).
}
\concept{dataset_erddap}
\keyword{datasets}
