% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciea_HMS.R
\docType{data}
\name{cciea_HMS}
\alias{cciea_HMS}
\title{Highly Migratory Species}
\format{
A data frame with 66 rows and 29 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{-5.99616E8, 1.4516064E9}}
\item{albacore_tuna_recruitment}{Albacore tuna recruitment () \link{126232.0, 307587.0}}
\item{albacore_tuna_spawning_stock_biomass}{Albacore tuna spawning stock biomass () \link{69427.6, 139481.0}}
\item{bigeye_tuna_recruitment}{Bigeye tuna recruitment () \link{9697.0, 39444.0}}
\item{bigeye_tuna_spawning_stock_biomass}{Bigeye tuna spawning stock biomass () \link{72327.0, 271970.0}}
\item{bluefin_tuna_recruitment}{Bluefin tuna recruitment () \link{3465.0, 34011.0}}
\item{bluefin_tuna_spawning_stock_biomass}{Bluefin tuna spawning stock biomass () \link{11445.0, 160005.0}}
\item{blue_marlin_recruitment}{Blue marlin recruitment () \link{589.3, 1181.18}}
\item{blue_marlin_spawning_stock_biomass}{Blue marlin spawning stock biomass () \link{20972.0, 71806.5}}
\item{skipjack_tuna_recruitment_relative_index}{Skipjack tuna recruitment index () \link{0.4760525, 2.503387}}
\item{skipjack_tuna_relative_biomass_index}{Skipjack tuna relative biomass index () \link{0.4711875, 2.244901}}
\item{swordfish_eastern_pacific_exploitable_total_biomass}{Eastern Pacific exploitable swordfish total biomass () \link{31510.0, 67070.0}}
\item{swordfish_western_central_pacific_exploitable_total_biomass}{Western Central Pacific exploitable swordfish total biomass () \link{60200.0, 121300.0}}
\item{yellowfin_tuna_recruitment}{Yellowfin tuna recruitment () \link{286403.0, 1319180.0}}
\item{yellowfin_tuna_spawning_stock_biomass_index}{Yellowfin tuna spawning stock biomass index () \link{2403.63, 9929.03}}
\item{albacore_tuna_recruitment_se}{Albacore tuna recruitment standard deviation () \link{38513.9, 120441.0}}
\item{albacore_tuna_spawning_stock_biomass_se}{Albacore tuna spawning stock biomass standard deviation () \link{27236.0, 55372.7}}
\item{bigeye_tuna_recruitment_se}{Bigeye tuna recruitment standard deviation () \link{2058.0, 8498.0}}
\item{bigeye_tuna_spawning_stock_biomass_se}{Bigeye tuna spawning stock biomass standard deviation () \link{23433.0, 76812.0}}
\item{bluefin_tuna_recruitment_se}{Bluefin tuna recruitment standard deviation () \link{620.0, 5302.0}}
\item{bluefin_tuna_spawning_stock_biomass_se}{Bluefin tuna spawning stock biomass standard deviation () \link{2337.0, 58187.0}}
\item{blue_marlin_recruitment_se}{Blue marlin recruitment standard deviation () \link{37.27, 221.6}}
\item{blue_marlin_spawning_stock_biomass_se}{Blue marlin spawning stock biomass standard deviation () \link{2576.58, 19780.7}}
\item{skipjack_tuna_recruitment_relative_index_se}{Skipjack tuna recruitment index standard deviation () \link{0.0, 0.0}}
\item{skipjack_tuna_relative_biomass_index_se}{Skipjack tuna relative biomass index standard deviation () \link{0.0, 0.0}}
\item{swordfish_eastern_pacific_exploitable_total_biomass_se}{Eastern Pacific exploitable swordfish total biomass standard deviation () \link{11050.0, 22520.0}}
\item{swordfish_western_central_pacific_exploitable_total_biomass_se}{Western Central Pacific exploitable swordfish total biomass standard deviation () \link{14810.0, 30900.0}}
\item{yellowfin_tuna_recruitment_se}{Yellowfin tuna recruitment standard deviation () \link{21845.0, 257967.0}}
\item{yellowfin_tuna_spawning_stock_biomass_index_se}{Yellowfin tuna spawning stock biomass index standard deviation () \link{128.5, 461.61}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_HMS/index.html}
}
\usage{
cciea_HMS
}
\description{
This index shows modeled spawning stock biomass of bigeye tuna from the latest stock assessment report, which was completed through the Inter-American Tropical Tuna Commission (IATTC), using Stock Synthesis V3. The assessment assumes that there is one stock of bigeye in the eastern Pacific. However, the assessment report acknowledges that recent tagging research suggests that bigeye undertake extensive longitudinal movements, which may be at odds with this assumption. The latest assessment report is available from www.iattc.org/Meetings/Meetings2017/JUL/PDFs/Docs/_English/IATTC-92-04a_The-fishery-in-2016-and-status-of-the-tuna-and-billfish-stocks.pdf; This index represents modeled spawning potential biomass from the latest (2017) stock assessment report, completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). North Pacific albacore are considered to be one stock throughout the North Pacific Ocean, although some studies have suggested that a northern and southern stock may be present within the assessment area. They are fished throughout their range by multiple countries, mostly using surface gear (troll and pole and line), as well as pelagic longlines and other gears. Their population dynamics are assessed using an age-, length- and sex-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; Skipjack tuna are assumed to comprise one contiguous stock throughout the Pacific ocean. In the eastern Pacific, they are primarily fished in tropical waters, using purse seine gear. Skipjack are difficult to assess with standard stock assessment methods, due to high and variable productivity, and uncertainties in natural mortality and growth. This species is thus assessed using a simple model which generates indicators of biomass, recruitment and exploitation rate, and compares these to historically observed values (Maunder & Deriso 2007). The stock assessment is completed through the Inter-American Tropical Tuna Commission (IATTC). The relative biomass index shown is from the latest update assessment, including data up to 2016, which is available from www.iattc.org/Meetings/Meetings2017/JUL/PDFs/Docs/_English/IATTC-92-04a_The-fishery-in-2016-and-status-of-the-tuna-and-billfish-stocks.pdf; Estimates of recruitment are derived from the assessment model. Skipjack tuna are assumed to comprise one contiguous stock throughout the Pacific ocean. In the eastern Pacific, they are primarily fished in tropical waters, using purse seine gear. Skipjack are difficult to assess with standard stock assessment methods, due to high and variable productivity, and uncertainties in natural mortality and growth. This species is thus assessed using a simple model which generates indicators of biomass, recruitment and exploitation rate, and compares these to historically observed values (Maunder & Deriso 2007). The stock assessment is completed through the Inter-American Tropical Tuna Commission (IATTC). The relative biomass index shown is from the latest update assessment, including data up to 2016, which is available from www.iattc.org/Meetings/Meetings2017/JUL/PDFs/Docs/_English/IATTC-92-04a_The-fishery-in-2016-and-status-of-the-tuna-and-billfish-stocks.pdf; Estimates of recruitment are derived from the assessment model. The 2017 stock assessment was an update of the 2016 full assessment. Yellowfin tuna are assessed through the Inter-American Tropical Tuna Commission (IATTC), using Stock Synthesis V3. They are assumed to comprise one stock throughout the Pacific, although tagging data suggest considerable regional fidelity. In the eastern Pacific, they are primarily fished in tropical waters, from Baja California south. The latest assessment report is available from www.iattc.org/Meetings/Meetings2017/JUL/PDFs/Docs/_English/IATTC-92-04a_The-fishery-in-2016-and-status-of-the-tuna-and-billfish-stocks.pdf; This series shows the modeled spawning stock biomass index from the 2017 stock assessment, which was an update of the 2016 full assessment. Yellowfin tuna are assessed through the Inter-American Tropical Tuna Commission (IATTC), using Stock Synthesis V3. They are assumed to comprise one stock throughout the Pacific, although tagging data suggest considerable regional fidelity. In the eastern Pacific, they are primarily fished in tropical waters, from Baja California south. The latest assessment report is available from www.iattc.org/Meetings/Meetings2017/JUL/PDFs/Docs/_English/IATTC-92-04a_The-fishery-in-2016-and-status-of-the-tuna-and-billfish-stocks.pdf; This index represents modeled spawning stock biomass from the latest (2016) stock assessment report, completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). Blue marlin are considered to be one stock throughout the Pacific Ocean, and the majority of catch is from pelagic longlines. Their population dynamics are assessed using an age-, length- and sex-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; Estimates of annual recruitment are derived from the stock assessment model. The modeled spawning stock biomass from the latest (2016) stock assessment report was completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). Blue marlin are considered to be one stock throughout the Pacific Ocean, and the majority of catch is from pelagic longlines. Their population dynamics are assessed using an age-, length- and sex-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; Annual recruitment is derived from the stock assessment model, and is primarily indexed by catches from troll fisheries on age-0 juvenile fish near Japan.  m the latest (2016) stock assessment rep was complted byleted through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). Pacific bluefin are considered to be one stock throughout the Pacific Ocean, and are fished throughout their range by multiple countries and fishing gears. At present, the majority are caught by purse seine gear. Their population dynamics are assessed using a fully integrated age-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html .; This index represents modeled spawning stock biomass from the latest (2016) stock assessment report, completed through the International Scientific Committee for Tuna and Tuna-like Species in the North cific Ocean (ISC). Pacific bluefin are considered to be one stock throughout the Pacific Ocean, and are fished throughout their range by multiple countries and fishing gears. At present, the majority are caught by purse seine gear. Their population dynamics are assessed using a fully integrated age-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; Estimates of annual recruitment are derived from the stock assessment model. The latest (2017) stock assessment report was completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). North Pacific albacore are considered to be one stock throughout the North Pacific Ocean, although some studies have suggested that a northern and southern stock may be present within the assessment area. They are fished throughout their range by multiple countries, mostly using surface gear (troll and pole and line), as well as pelagic longlines and other gears. Their population dynamics are assessed using an age-, length- and sex-structured model (Stock Synthesis v3). The full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html; These indices represent modeled exploitable biomass for the western central and eastern Pacific swordfish stocks, from the latest (2014) stock assessment report, completed through the International Scientific Committee for Tuna and Tuna-like Species in the North Pacific Ocean (ISC). Swordfish are considered to comprise two stocks in the North Pacific. The western and central Pacific stock is located throughout the entire North Pacific, except for off Baja California, and central and south America, which is occupied by the eastern Pacific stock. The highest catches are from pelagic longline gears. The population dynamics of both stocks are assessed using a Bayesian state-space generalized surplus production model. The  full assessment is available from http://isc.fra.go.jp/reports/stock_assessments.html.; Estimates of recruitment are derived from the assessment model. The latest stock assessment report  was completed through the Inter-American Tropical Tuna Commission (IATTC), using Stock Synthesis V3. The assessment assumes that there is one stock of bigeye in the eastern Pacific. However, the assessment report acknowledges that recent tagging research suggests that bigeye undertake extensive longitudinal movements, which may be at odds with this assumption. The latest assessment report is available from www.iattc.org/Meetings/Meetings2017/JUL/PDFs/Docs/_English/IATTC-92-04a_The-fishery-in-2016-and-status-of-the-tuna-and-billfish-stocks.pdf
}
\concept{dataset_erddap}
\keyword{datasets}
