% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_MM_HB_REL}
\alias{cciea_MM_HB_REL}
\title{Relative humpback whale abundance}
\format{
A data frame with 42 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [8.204544E8, 1.5463008E9]}
\item{humpback_rate}{Relative humpback abundance (index) [0.0, 17.2]}
\item{humpback_anomaly}{Relative humpback abundance anomaly (index) [-6.3381, 10.8619]}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_MM_HB_REL/index.html}
}
\usage{
cciea_MM_HB_REL
}
\description{
Regional estimate of Relative humpback whale abundance values from surveys conducted by the Farallon Institute OR the RREAS surveys (or combination).
}
\concept{dataset_erddap}
\keyword{datasets}
