% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_SL3}
\alias{cciea_OC_SL3}
\title{Monthly Sea Level: San Diego CA (32.7N 117.2W)}
\format{
A data frame with 6571 rows and 5 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) [-2.0184768E9, 1.4501376E9]}
\item{sea_level}{Sea Level (mm) [1766.71, 2282.7]}
\item{std_error}{Standard Error (mm) [0.0, 0.0]}
\item{station}{ () []}
\item{latitude}{Latitude (degrees_north) [32.7, 32.7]}
\item{longitude}{Longitude (degrees_east) [117.2, 117.2]}
\item{depth}{Depth (m) []}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_SL3/index.html}
}
\usage{
cciea_OC_SL3
}
\description{
The California Current Large Marine Ecosystem (CCLME) is primarily driven by bottom-up physical oceanographic processes, thus understanding trends in the physical state can inform our knowledge of ecosystem processes and management of ecosystem services.
}
\concept{dataset_erddap}
\keyword{datasets}
