% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_erddap.R
\docType{data}
\name{cciea_OC_VWND2}
\alias{cciea_OC_VWND2}
\title{Meridional Wind Buoy 46014 (39.2N 124.0W)}
\format{
A data frame with 439 rows and 2 variables:
\describe{
\item{time}{Time (seconds since 1970-01-01T00:00:00Z) \link{3.549312E8, 1.6121376E9}}
\item{vwnd}{Wind Speed, Meridional (m s-1) \link{-7.22959, 6.46426}}
}
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/info/cciea_OC_VWND2/index.html}
}
\usage{
cciea_OC_VWND2
}
\description{
The National Data Buoy Center (NDBC) distributes meteorological data from moored buoys maintained by NDBC and others. The data is from NOAA NDBC. It has been reformatted by NOAA Coastwatch, West Coast Node and then monthly averaged. This dataset has both historical data and near real time data.
}
\concept{dataset_erddap}
\keyword{datasets}
