% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeTimeSeries.R
\name{GetNorm}
\alias{GetNorm}
\title{Normalize Time Series, original}
\usage{
GetNorm(X, Y, Name)
}
\arguments{
\item{X}{time as year, year-month or year-month-day}

\item{Y}{value to plot over time}

\item{Name}{title of plot}
}
\value{

}
\description{
This function will normalize multiple time series and produce the quadplot.
Your data need three columns: year, index, time series name.
}
\details{
Example command: QuadPlot(MyData$year, MyData$index, MyData$timeseries)
file made: 2018-10-23
}
\concept{plot_original}
